Moved contents of improver.utilities.cube_metadata to logical places within the metadata subpackage.  Split out unit tests and updated imports.  No functional changes.
