+++ lib/improver/cli/threshold.py
@@ -38,8 +38,8 @@ import numpy as np
 
 from improver.argparser import ArgParser
 from improver.blending.calculate_weights_and_blend import WeightAndBlend
+from improver.metadata.probabilistic import in_vicinity_name_format
 from improver.threshold import BasicThreshold
-from improver.utilities.cube_metadata import in_vicinity_name_format
 from improver.utilities.load import load_cube
 from improver.utilities.save import save_netcdf
 from improver.utilities.spatial import OccurrenceWithinVicinity
