+++ lib/improver/cli/update_grid_metadata.py
@@ -37,7 +37,7 @@ version 1.1.0 to StaGE version 1.2.0.
 import os
 
 from improver.argparser import ArgParser
-from improver.utilities.cube_metadata import update_stage_v110_metadata
+from improver.metadata.amend import update_stage_v110_metadata
 from improver.utilities.load import load_cube
 from improver.utilities.save import save_netcdf
