+++ lib/improver/spotdata/neighbour_finding.py
@@ -37,9 +37,9 @@ import cartopy.crs as ccrs
 import numpy as np
 from scipy.spatial import cKDTree
 
+from improver.metadata.utilities import create_coordinate_hash
 from improver.spotdata.build_spotdata_cube import build_spotdata_cube
 from improver.utilities.cube_manipulation import enforce_coordinate_ordering
-from improver.utilities.cube_metadata import create_coordinate_hash
 
 
 class NeighbourSelection:
