+++ lib/improver/tests/blending/blend_across_adjacent_points/test_TriangularWeightedBlendAcrossAdjacentPoints.py
@@ -40,9 +40,9 @@ from iris.tests import IrisTest
 
 from improver.blending.blend_across_adjacent_points import \
     TriangularWeightedBlendAcrossAdjacentPoints
+from improver.metadata.amend import add_coord
 from improver.tests.set_up_test_cubes import set_up_variable_cube
 from improver.utilities.cube_manipulation import concatenate_cubes
-from improver.utilities.cube_metadata import add_coord
 from improver.utilities.warnings_handler import ManageWarnings
