+++ lib/improver/tests/blending/weights/test_WeightsUtilities.py
@@ -38,8 +38,7 @@ import numpy as np
 from iris.tests import IrisTest
 
 from improver.blending.weights import WeightsUtilities
-from improver.tests.utilities.test_cube_metadata import (
-    create_cube_with_threshold)
+from improver.tests.metadata.test_amend import create_cube_with_threshold
 from improver.utilities.cube_checker import find_threshold_coordinate
