+++ lib/improver/tests/ensemble_calibration/ensemble_calibration/helper_functions.py
@@ -41,8 +41,8 @@ from iris.cube import Cube
 from iris.tests import IrisTest
 import numpy as np
 
+from improver.metadata.probabilistic import extract_diagnostic_name
 from improver.tests.set_up_test_cubes import set_up_variable_cube
-from improver.utilities.cube_metadata import extract_diagnostic_name
 from improver.utilities.warnings_handler import ManageWarnings
 
 IGNORED_MESSAGES = ["Collapsing a non-contiguous coordinate"]
