+++ lib/improver/tests/spotdata/spotdata/test_SpotExtraction.py
@@ -36,9 +36,9 @@ import iris
 import numpy as np
 from iris.tests import IrisTest
 
+from improver.metadata.utilities import create_coordinate_hash
 from improver.spotdata.build_spotdata_cube import build_spotdata_cube
 from improver.spotdata.spot_extraction import SpotExtraction
-from improver.utilities.cube_metadata import create_coordinate_hash
 
 
 class Test_SpotExtraction(IrisTest):
