+++ lib/improver/tests/spotdata/spotdata/test_check_grid_match.py
@@ -35,10 +35,10 @@ import unittest
 import numpy as np
 from iris.tests import IrisTest
 
+from improver.metadata.utilities import create_coordinate_hash
 from improver.spotdata.build_spotdata_cube import build_spotdata_cube
 from improver.spotdata.spot_extraction import check_grid_match
 from improver.tests.set_up_test_cubes import set_up_variable_cube
-from improver.utilities.cube_metadata import create_coordinate_hash
 
 
 class Test_check_grid_match(IrisTest):
